+++ src/python_bindings/bindings/boolean_function.cpp
@@ -220,6 +220,13 @@ namespace hal
             :rtype: hal_py.BooleanFunction
         )");
 
+        py_boolean_function.def("optimize_constants", &BooleanFunction::optimize_constants, R"(
+            Removes constant values whenever possible.
+
+            :returns: The optimized Boolean function.
+            :rtype: hal_py.BooleanFunction
+        )");
+
         py_boolean_function.def("get_truth_table", &BooleanFunction::get_truth_table, py::arg("ordered_variables") = std::vector<std::string>(), py::arg("remove_unknown_variables") = false, R"(
             Get the truth table outputs of the Boolean function.